+++ src/main/java/ti4/discord/interactions/buttons/handlers/faction/base/arborec/ArborecButtonHandlers.java
@@ -1,9 +1,11 @@
 package ti4.discord.interactions.buttons.handlers.faction.base.arborec;
 
+import java.util.ArrayList;
 import java.util.List;
 import lombok.experimental.UtilityClass;
 import net.dv8tion.jda.api.components.buttons.Button;
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
+import ti4.discord.interactions.buttons.Buttons;
 import ti4.discord.interactions.routing.ButtonHandler;
 import ti4.game.Game;
 import ti4.game.Player;
@@ -11,13 +13,15 @@ import ti4.game.Tile;
 import ti4.helpers.ButtonHelper;
 import ti4.helpers.ButtonHelperAgents;
 import ti4.helpers.Helper;
+import ti4.helpers.Units.UnitType;
 import ti4.image.TileHelper;
 import ti4.message.MessageHelper;
 import ti4.service.emoji.UnitEmojis;
 import ti4.service.unit.AddUnitService;
+import ti4.service.unit.CheckUnitContainmentService;
 
 @UtilityClass
-class ArborecButtonHandlers {
+public class ArborecButtonHandlers {
 
     @ButtonHandler("arboAgentOn_")
     public static void arboAgentOn(ButtonInteractionEvent event, Player player, String buttonID, Game game) {
@@ -93,5 +97,32 @@ class ArborecButtonHandlers {
                 + ButtonHelper.getListOfStuffAvailableToSpend(player, game);
         MessageHelper.sendMessageToChannelWithButtons(event.getChannel(), message, buttons);
         ButtonHelper.deleteMessage(event);
+    }
+
+    public static List<Button> getArboHeroButtons(Game game, Player player) {
+        List<Button> buttons = new ArrayList<>();
+        List<Tile> tiles = new ArrayList<>();
+        tiles.addAll(CheckUnitContainmentService.getTilesContainingPlayersUnits(game, player, UnitType.Infantry));
+        tiles.addAll(CheckUnitContainmentService.getTilesContainingPlayersUnits(game, player, UnitType.Mech));
+        List<String> poses = new ArrayList<>();
+        for (Tile tile : tiles) {
+            if (!poses.contains(tile.getPosition())) {
+                buttons.add(Buttons.green(
+                        "arboHeroBuild_" + tile.getPosition(), tile.getRepresentationForButtons(game, player)));
+                poses.add(tile.getPosition());
+            }
+        }
+        buttons.add(Buttons.red("deleteButtons", "Done"));
+        return buttons;
+    }
+
+    @ButtonHandler("arboHeroBuild_")
+    public static void resolveArboHeroBuild(Game game, Player player, ButtonInteractionEvent event, String buttonID) {
+        String pos = buttonID.split("_")[1];
+        List<Button> buttons =
+                Helper.getPlaceUnitButtons(event, player, game, game.getTileByPosition(pos), "arboHeroBuild", "place");
+        String message = player.getRepresentation() + " Use the buttons to produce units. ";
+        MessageHelper.sendMessageToChannelWithButtons(player.getCorrectChannel(), message, buttons);
+        ButtonHelper.deleteButtonAndDeleteMessageIfEmpty(event);
     }
 }
